Final pre-commit tests and linting verified
