+++ src/coreason_manifest/core/workflow/nodes/__init__.py
@@ -1,0 +1,35 @@
+# Prosperity-3.0
+from typing import Any
+
+from coreason_manifest.core.primitives.registry import resolve_node_union
+
+from .agent import AgentNode, CognitiveProfile
+from .base import Constraint, ConstraintOperator, LockConfig, Node
+from .human import HumanNode, SteeringConfig
+from .oversight import EmergenceInspectorNode, InspectorNode, InspectorNodeBase
+from .planner import PlannerNode
+from .routing import SwitchNode
+from .swarm import SwarmNode
+from .system import PlaceholderNode
+
+# AnyNode is now resolved dynamically
+AnyNode: Any = resolve_node_union()
+
+__all__ = [
+    "AgentNode",
+    "AnyNode",
+    "CognitiveProfile",
+    "Constraint",
+    "ConstraintOperator",
+    "EmergenceInspectorNode",
+    "HumanNode",
+    "InspectorNode",
+    "InspectorNodeBase",
+    "LockConfig",
+    "Node",
+    "PlaceholderNode",
+    "PlannerNode",
+    "SteeringConfig",
+    "SwarmNode",
+    "SwitchNode",
+]
